adjust api for RSA methods
