+++ Tests/BLCryptoTests/EncryptDecryptTests.swift
@@ -17,8 +17,8 @@ class EncryptDecryptTests: XCTestCase {
         let str = "Clear Text"
         let clearMessage = try ClearMessage(string: str, using: .utf8)
         
-        let encrypted = try clearMessage.encrypted(with: publicKey, paddingType: .pkcs1)
-        let decrypted = try encrypted.decrypted(with: privateKey, paddingType: .pkcs1)
+        let encrypted = try RSA.encrypt(clearMessage, with: publicKey, paddingType: .pkcs1)
+        let decrypted = try RSA.decrypt(encrypted, with: privateKey, paddingType: .pkcs1)
         
         XCTAssertEqual(try? decrypted.string(encoding: .utf8), str)
     }
@@ -27,8 +27,8 @@ class EncryptDecryptTests: XCTestCase {
         let str = [String](repeating: "a", count: 9999).joined(separator: "")
         let clearMessage = try ClearMessage(string: str, using: .utf8)
         
-        let encrypted = try clearMessage.encrypted(with: publicKey, paddingType: .pkcs1)
-        let decrypted = try encrypted.decrypted(with: privateKey, paddingType: .pkcs1)
+        let encrypted = try RSA.encrypt(clearMessage, with: publicKey, paddingType: .pkcs1)
+        let decrypted = try RSA.decrypt(encrypted, with: privateKey, paddingType: .pkcs1)
         
         XCTAssertEqual(try? decrypted.string(encoding: .utf8), str)
     }
@@ -37,8 +37,8 @@ class EncryptDecryptTests: XCTestCase {
         let data = TestUtils.randomData(count: 2048)
         let clearMessage = ClearMessage(data: data)
         
-        let encrypted = try clearMessage.encrypted(with: publicKey, paddingType: .pkcs1)
-        let decrypted = try encrypted.decrypted(with: privateKey, paddingType: .pkcs1)
+        let encrypted = try RSA.encrypt(clearMessage, with: publicKey, paddingType: .pkcs1)
+        let decrypted = try RSA.decrypt(encrypted, with: privateKey, paddingType: .pkcs1)
         
         XCTAssertEqual(decrypted.data, data)
     }
@@ -48,7 +48,7 @@ class EncryptDecryptTests: XCTestCase {
 //
 //        let data = TestUtils.randomData(count: 128)
 //        let clearMessage = ClearMessage(data: data)
-//        let encrypted = try clearMessage.encrypted(with: publicKey, padding: [])
+//        let encrypted = try RSA.encrypt(clearMessage, with: publicKey, padding: [])
 //
 //        let clearMessage2 = ClearMessage(data: encrypted.data)
 //        let encrypted2 = try clearMessage2.encrypted(with: publicKey, padding: [])
@@ -56,7 +56,7 @@ class EncryptDecryptTests: XCTestCase {
 //        XCTAssertEqual(data.count, encrypted.data.count)
 //        XCTAssertEqual(data.count, encrypted2.data.count)
 //
-//        let decrypted = try encrypted.decrypted(with: privateKey, padding: [])
+//        let decrypted = try RSA.decrypt(encrypted, with: privateKey, padding: [])
 //
 //        XCTAssertEqual(decrypted.data, data)
 //    }
@@ -65,8 +65,8 @@ class EncryptDecryptTests: XCTestCase {
         let data = TestUtils.randomData(count: 2048)
         let clearMessage = ClearMessage(data: data)
         
-        let encrypted = try clearMessage.encrypted(with: publicKey, paddingType: .oaep)
-        let decrypted = try encrypted.decrypted(with: privateKey, paddingType: .oaep)
+        let encrypted = try RSA.encrypt(clearMessage, with: publicKey, paddingType: .oaep)
+        let decrypted = try RSA.decrypt(encrypted, with: privateKey, paddingType: .oaep)
         
         XCTAssertEqual(decrypted.data, data)
     }
@@ -80,29 +80,29 @@ class EncryptDecryptTests: XCTestCase {
         
         // Encrypt with old public key, decrypt with old private key
         do {
-            let encrypted = try clearMessage.encrypted(with: publicKey, paddingType: .pkcs1)
-            let decrypted = try encrypted.decrypted(with: privateKey, paddingType: .pkcs1)
+            let encrypted = try RSA.encrypt(clearMessage, with: publicKey, paddingType: .pkcs1)
+            let decrypted = try RSA.decrypt(encrypted, with: privateKey, paddingType: .pkcs1)
             XCTAssertEqual(decrypted.data, data)
         }
         
         // Encrypt with old public key, decrypt with new private key
         do {
-            let encrypted = try clearMessage.encrypted(with: publicKey, paddingType: .pkcs1)
-            let decrypted = try encrypted.decrypted(with: newPrivateKey, paddingType: .pkcs1)
+            let encrypted = try RSA.encrypt(clearMessage, with: publicKey, paddingType: .pkcs1)
+            let decrypted = try RSA.decrypt(encrypted, with: newPrivateKey, paddingType: .pkcs1)
             XCTAssertEqual(decrypted.data, data)
         }
         
         // Encrypt with new public key, decrypt with old private key
         do {
-            let encrypted = try clearMessage.encrypted(with: newPublicKey, paddingType: .pkcs1)
-            let decrypted = try encrypted.decrypted(with: privateKey, paddingType: .pkcs1)
+            let encrypted = try RSA.encrypt(clearMessage, with: newPublicKey, paddingType: .pkcs1)
+            let decrypted = try RSA.decrypt(encrypted, with: privateKey, paddingType: .pkcs1)
             XCTAssertEqual(decrypted.data, data)
         }
         
         // Encrypt with new public key, decrypt with new private key
         do {
-            let encrypted = try clearMessage.encrypted(with: newPublicKey, paddingType: .pkcs1)
-            let decrypted = try encrypted.decrypted(with: newPrivateKey, paddingType: .pkcs1)
+            let encrypted = try RSA.encrypt(clearMessage, with: newPublicKey, paddingType: .pkcs1)
+            let decrypted = try RSA.decrypt(encrypted, with: newPrivateKey, paddingType: .pkcs1)
             XCTAssertEqual(decrypted.data, data)
         }
     }
@@ -117,29 +117,29 @@ class EncryptDecryptTests: XCTestCase {
         
         // Encrypt with old public key, decrypt with old private key
         do {
-            let encrypted = try clearMessage.encrypted(with: publicKey, paddingType: .pkcs1)
-            let decrypted = try encrypted.decrypted(with: privateKey, paddingType: .pkcs1)
+            let encrypted = try RSA.encrypt(clearMessage, with: publicKey, paddingType: .pkcs1)
+            let decrypted = try RSA.decrypt(encrypted, with: privateKey, paddingType: .pkcs1)
             XCTAssertEqual(decrypted.data, data)
         }
         
         // Encrypt with old public key, decrypt with new private key
         do {
-            let encrypted = try clearMessage.encrypted(with: publicKey, paddingType: .pkcs1)
-            let decrypted = try encrypted.decrypted(with: newPrivateKey, paddingType: .pkcs1)
+            let encrypted = try RSA.encrypt(clearMessage, with: publicKey, paddingType: .pkcs1)
+            let decrypted = try RSA.decrypt(encrypted, with: newPrivateKey, paddingType: .pkcs1)
             XCTAssertEqual(decrypted.data, data)
         }
         
         // Encrypt with new public key, decrypt with old private key
         do {
-            let encrypted = try clearMessage.encrypted(with: newPublicKey, paddingType: .pkcs1)
-            let decrypted = try encrypted.decrypted(with: privateKey, paddingType: .pkcs1)
+            let encrypted = try RSA.encrypt(clearMessage, with: newPublicKey, paddingType: .pkcs1)
+            let decrypted = try RSA.decrypt(encrypted, with: privateKey, paddingType: .pkcs1)
             XCTAssertEqual(decrypted.data, data)
         }
         
         // Encrypt with new public key, decrypt with new private key
         do {
-            let encrypted = try clearMessage.encrypted(with: newPublicKey, paddingType: .pkcs1)
-            let decrypted = try encrypted.decrypted(with: newPrivateKey, paddingType: .pkcs1)
+            let encrypted = try RSA.encrypt(clearMessage, with: newPublicKey, paddingType: .pkcs1)
+            let decrypted = try RSA.decrypt(encrypted, with: newPrivateKey, paddingType: .pkcs1)
             XCTAssertEqual(decrypted.data, data)
         }
     }
